Extract hourly time series from pre-cut CHELSA V2.1 raster bricks

Function extracts precipitation (pr) or temperature (tas) data from raster bricks and prepares a dataframe for later import in RSMinerve. Use write.table() function to export as csv.

generate_DBcsv_from_CHELSA21nc(
  CHELSA21_dir,
  data_type,
  HRU,
  HRU_name_column,
  start_year,
  end_year,
  tz = "UTC"
)

Arguments

CHELSA21_dir
Path to CHELSA precipitation or temperature nc files which have been clipped to the bounding box of the catchment of interest.

data_type
Specifies the variable to extract. Precipitation data : data_type = "pr" or temperature data: data_type = "tas".

HRU
Vector object (multi-polygons) with hydrological response units for which the CHELSA V2.1 data should be extracted.

HRU_name_column
(string) Name of the column in HRU with unique names for the hydrological response units. These names will be used as identifiers in RSMinerve.

start_year
First year for which data should be made available (assuming data 'is' available from the start of that year)

end_year
Last year from which data should be extracted (assuming data 'is' actually available until the end of that year)

tz
Time zone string to be passed to as_datetime. Defaults to "UTC".

Value

Dataframe tibble with temperature in deg. C. and/or precipitation in mm/h

Details

The CHELSA V2.1 raster bricks should be clipped to the bounding box of the catchment to be modelled for higher speed. The clipping can be done efficiently in cda (to do: Write a vignette on how to use cda for this). The CHELSA V2.1 nc files contain precipitation or temperature data for each raster point and each day of one year between 1979 and 2018. In some areas of Central Asia, the quality of the ERA5 product underlying the CHELSA v2.1 data set is not sufficient and only data until 2011 should be used. Please validate the CHELSA V2.1 data set prior to using it for modelling. The function assumes GMT time zone.

Note

Acknowledgement: This function builds on hydrosolutions/riversCentralAsia::generate_ERA5_Subbasin_CSV