Reads RS Minerve result file and returns data.

Reads RS Minerve result file and returns data.

readResultCSV(filename, tz = "UTC")

Arguments

filename

Path to file to be read. It is any csv file that is written by saving data from the Selection and plots tab in RSMinerve to a csv file. Depending on the systems locale, the csv may use comas or semi-colons as separators. The decimal separator must be a point, commas are not recognized.

tz

Time zone character passed to as.POSIXct. Defaults to "UTC".

Value

data Tibble with data from file. Null if file not found or cant be read.

Note

An example RSMinerve results csv file is available under http://raw.githubusercontent.com/hydrosolutions/RSMinerveR/main/tests/testthat/test_translateCSVtoDST.csv. If the function generates funny numbers, please check the csv file. The decimal separators must be . and not ,. If you find commas please change the settings of your PC (Control Panel, Clock and Region, Region, Additional settings, select . for the Decimal symbol, press "Apply" and close the Control Panel windows. and RSMinerve will write ,.